tests/wpt/web-platform-tests/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/002.html

<!doctype html>
<title>document.open during parsing</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
 var log = document.getElementById("log");
 assert_equals(document.open(), document);
 assert_equals(document.getElementById("log"), log);
})
</script>
